# backend/openapi/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Women Couture API",
    "version": "0.1.0",
    "description": "REST API for Women Couture"
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/v1"
    }
  ],
  "components": {
    "schemas": {
...
      "ProductBase": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "type": {
            "type": "string",
            "enum": [
              "UNSTITCHED",
              "THREE_PC",
              "TWO_PC",
              "SEPARATE"
            ]
          },
          "isActive": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "type",
          "isActive",
          "createdAt",
          "updatedAt"
        ]
      },
      "ProductImage": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "url": {
            "type": "string"
          },
          "order": {
            "type": "integer"
          }
        },
        "required": [
          "id",
          "url",
          "order"
        ]
      },
      "ProductVariant": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "productId": {
            "type": "string"
          },
          "color": {
            "type": "string"
          },
          "sku": {
            "type": "string",
            "nullable": true
          },
          "salePrice": {
            "type": "number",
            "nullable": true
          },
          "purchasePrice": {
            "type": "number",
            "nullable": true
          },
          "stockQty": {
            "type": "integer"
          },
          "images": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductImage"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "productId",
          "color",
          "sku",
          "salePrice",
          "purchasePrice",
          "stockQty",
          "createdAt",
          "updatedAt"
        ]
      },
      "ProductWithVariants": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ProductBase"
          },
          {
            "type": "object",
            "properties": {
              "variants": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/ProductVariant"
                }
              }
            },
            "required": [
              "variants"
            ]
          }
        ]
      },
      "CreateProductBody": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1
          },
          "description": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "UNSTITCHED",
              "THREE_PC",
              "TWO_PC",
              "SEPARATE"
            ]
          },
          "variants": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "color": {
                  "type": "string",
                  "minLength": 1
                },
                "sku": {
                  "type": "string"
                },
                "salePrice": {
                  "type": "number",
                  "minimum": 0,
                  "exclusiveMinimum": true
                },
                "purchasePrice": {
                  "type": "number",
                  "minimum": 0,
                  "exclusiveMinimum": true
                }
              },
              "required": [
                "color",
                "salePrice"
              ]
            }
          }
        },
        "required": [
          "name",
          "type"
        ]
      },
...
      "UpdateProductBody": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1
          },
          "description": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "UNSTITCHED",
              "THREE_PC",
              "TWO_PC",
              "SEPARATE"
            ]
          },
          "isActive": {
            "type": "boolean"
          }
        }
      },
...
    },
...
  },
  "paths": {
...
    "/admin/products": {
      "get": {
        "summary": "List all products",
        "tags": [
          "Products"
        ],
        "parameters": [
          {
            "schema": {
              "type": "number",
              "nullable": true,
              "description": "Number of records to skip"
            },
            "required": false,
            "description": "Number of records to skip",
            "name": "skip",
            "in": "query"
          },
          {
            "schema": {
              "type": "number",
              "nullable": true,
              "description": "Number of records to return"
            },
            "required": false,
            "description": "Number of records to return",
            "name": "take",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false"
              ],
              "description": "Filter by active state"
            },
            "required": false,
            "description": "Filter by active state",
            "name": "isActive",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductWithVariants"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new base product",
        "tags": [
          "Products"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductWithVariants"
                }
              }
            }
          }
        }
      }
    },
    "/admin/products/{productId}": {
      "get": {
        "summary": "Get a product by ID",
        "tags": [
          "Products"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "productId",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductWithVariants"
                }
              }
            }
          },
          "404": {
            "description": "Product not found"
          }
        }
      },
      "patch": {
        "tags": [
          "Products"
        ],
        "summary": "Update product details or status",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "productId",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductBase"
                }
              }
            }
          }
        }
      }
    },
...
  }
}